Very basic support for Format 3
Format 3 includes a string in the header indicating the default code
page. It also "supports custom serialization of ALTREP framework
objects" but I don't know what these are.

In the future it would be nice to use iconv to recode everything to
UTF-8, but for now just throw away the encoding string.